Move duplicate useInterval code into a separate file and import
Add isCancelled to missing useIntervals

Set loading status to not loading if cancelled
